Added sorting and tests
